Refactors compare-properties.ts, breaking it out into files of separate concerns
