+++ packages/graphql/src/schema/resolvers/subscriptions/subscribe.ts
@@ -24,9 +24,9 @@ import type Node from "../../../classes/Node";
 import type { NodeSubscriptionsEvent, RelationshipSubscriptionsEvent, SubscriptionsEvent } from "../../../types";
 import { filterAsyncIterator } from "./filter-async-iterator";
 import { SubscriptionAuth } from "./subscription-auth";
-import type { SubscriptionContext } from "./types";
+import type { EventType, SubscriptionContext } from "./types";
 import { updateDiffFilter } from "./update-diff-filter";
-import { subscriptionWhere } from "./where";
+import { subscriptionWhere } from "./where/where";
 
 export function subscriptionResolve(payload: [SubscriptionsEvent]): SubscriptionsEvent {
     if (!payload) {
@@ -46,7 +46,7 @@ export function generateSubscribeMethod({
     relationshipFields,
 }: {
     node: Node;
-    type: "create" | "update" | "delete" | "create_relationship" | "delete_relationship";
+    type: EventType;
     nodes?: Node[];
     relationshipFields?: Map<string, ObjectFields>;
 }) {
