+++ packages/graphql/src/schema/resolvers/subscriptions/types.ts
@@ -20,6 +20,8 @@
 import type { Neo4jGraphQLSubscriptionsPlugin } from "../../../types";
 import type { JwtPayload } from "../../../types/deprecated/auth/jwt-payload";
 
+export type EventType = "create" | "update" | "delete" | "create_relationship" | "delete_relationship";
+
 export type SubscriptionContext = {
     plugin: Neo4jGraphQLSubscriptionsPlugin;
     jwt?: JwtPayload;
@@ -31,3 +33,10 @@ export type SubscriptionConnectionContext = {
     };
     jwt?: JwtPayload;
 };
+
+export type StandardType = Record<string, Record<string, unknown>>;
+export type UnionType = Record<string, StandardType>;
+export type InterfaceType = Record<string, unknown | InterfaceSpecificType>;
+export type InterfaceSpecificType = Record<string, Record<string, unknown>>;
+export type RecordType = Record<string, unknown>;
+export type RelationshipType = Record<string, Record<string, UnionType | InterfaceType | StandardType>>;
