+++ packages/graphql/src/schema/resolvers/subscriptions/where/utils/parse-filter-property.ts
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) "Neo4j"
+ * Neo4j Sweden AB [http://neo4j.com]
+ *
+ * This file is part of Neo4j.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import type { WhereRegexGroups } from "../../../../../translate/where/utils";
+import { whereRegEx } from "../../../../../translate/where/utils";
+
+export function parseFilterProperty(key: string): { fieldName: string; operator: string | undefined } {
+    const match = whereRegEx.exec(key);
+    if (!match) {
+        throw new Error(`Failed to match key in filter: ${key}`);
+    }
+    const { fieldName, operator } = match.groups as WhereRegexGroups;
+    if (!fieldName) {
+        throw new Error(`Failed to find field name in filter: ${key}`);
+    }
+    return { fieldName, operator };
+}
